# docs/api/internal.openapi.yaml
openapi: 3.1.0
info:
  title: DeepSight Internal Admin API
  version: "1.0.0"
  description: |
    엔진스튜디오 내부 관리자 API — 페르소나 관리 · 매칭 랩 · 보안 · 운영

    > **주의**: 이 API는 내부 운영 도구입니다. 세션 기반 인증이 필요합니다.

  contact:
    name: DeepSight Internal Team
servers:
  - url: https://engine.deepsight.ai/api/internal
    description: Production (engine-studio)
...
paths:
...
  /personas/{id}/autonomy:
    get:
      operationId: getAutonomyPolicy
      tags: [Personas]
      summary: 자율 동작 정책 조회
      description: per-persona AutonomyPolicy 조회. 미설정 시 기본값 반환.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: 정책 조회 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutonomyPolicyResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateAutonomyPolicy
      tags: [Personas]
      summary: 자율 동작 정책 업데이트
      description: 부분 업데이트 지원. 기존 정책과 머지 후 검증.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutonomyPolicyInput"
      responses:
        "200":
          description: 정책 업데이트 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutonomyPolicyResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # -----------------------------------------------------------
  # Autonomy (자율 동작)
  # -----------------------------------------------------------
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, enum: [false] }
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
...
    AutonomyPolicyResponse:
      type: object
      properties:
        success: { type: boolean, enum: [true] }
        data:
          type: object
          properties:
            personaId: { type: string }
            personaName: { type: string }
            isCustom: { type: boolean }
            policy:
              $ref: "#/components/schemas/AutonomyPolicy"
    AutonomyPolicy:
      type: object
      properties:
        autoCorrection: { type: boolean }
        autoMemoryManagement: { type: boolean }
        metaCognitionEnabled: { type: boolean }
        correctionConfig:
          type: object
          properties:
            maxAutoSeverity: { type: string, enum: [minor, major] }
            minConfidence: { type: number, minimum: 0.7, maximum: 1.0 }
            dailyLimit: { type: integer, minimum: 1, maximum: 10 }
        memoryConfig:
          type: object
          properties:
            pruneConfidenceThreshold: { type: number, minimum: 0.0, maximum: 0.5 }
            maxPerCategory: { type: integer, minimum: 10, maximum: 1000 }
    AutonomyPolicyInput:
      type: object
      properties:
        autoCorrection: { type: boolean }
        autoMemoryManagement: { type: boolean }
        metaCognitionEnabled: { type: boolean }
        correctionConfig:
          type: object
          properties:
            maxAutoSeverity: { type: string, enum: [minor, major] }
            minConfidence: { type: number, minimum: 0.7, maximum: 1.0 }
            dailyLimit: { type: integer, minimum: 1, maximum: 10 }
        memoryConfig:
          type: object
          properties:
            pruneConfidenceThreshold: { type: number, minimum: 0.0, maximum: 0.5 }
            maxPerCategory: { type: integer, minimum: 10, maximum: 1000 }

    # Arena
...
  responses:
    Unauthorized:
      description: 인증 실패
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Authentication required
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: 리소스 없음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"